+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Redesign
 
 - Tables redesign
+- Redesigned friends status: you can now click on individual links instead of one big link
 
 # [2.0.0-alpha.9.2]